+++ pkg/services/authz/zanzana/logger/logger.go
@@ -1,4 +1,4 @@
-package zanzana
+package logger
 
 import (
 	"context"
@@ -8,13 +8,13 @@ import (
 	"github.com/grafana/grafana/pkg/infra/log"
 )
 
-// zanzanaLogger is a grafana logger wrapper compatible with OpenFGA logger interface
-type zanzanaLogger struct {
+// ZanzanaLogger is a grafana logger wrapper compatible with OpenFGA logger interface
+type ZanzanaLogger struct {
 	logger log.Logger
 }
 
-func newZanzanaLogger(logger log.Logger) *zanzanaLogger {
-	return &zanzanaLogger{
+func New(logger log.Logger) *ZanzanaLogger {
+	return &ZanzanaLogger{
 		logger: logger,
 	}
 }
@@ -36,50 +36,50 @@ func zapFieldsToArgs(fields []zap.Field) []any {
 	return args
 }
 
-func (l *zanzanaLogger) Debug(msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) Debug(msg string, fields ...zap.Field) {
 	l.logger.Debug(msg, zapFieldsToArgs(fields)...)
 }
 
-func (l *zanzanaLogger) Info(msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) Info(msg string, fields ...zap.Field) {
 	l.logger.Info(msg, zapFieldsToArgs(fields)...)
 }
 
-func (l *zanzanaLogger) Warn(msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) Warn(msg string, fields ...zap.Field) {
 	l.logger.Warn(msg, zapFieldsToArgs(fields)...)
 }
 
-func (l *zanzanaLogger) Error(msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) Error(msg string, fields ...zap.Field) {
 	l.logger.Error(msg, zapFieldsToArgs(fields)...)
 }
 
-func (l *zanzanaLogger) Panic(msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) Panic(msg string, fields ...zap.Field) {
 	l.logger.Error(msg, zapFieldsToArgs(fields)...)
 }
 
-func (l *zanzanaLogger) Fatal(msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) Fatal(msg string, fields ...zap.Field) {
 	l.logger.Error(msg, zapFieldsToArgs(fields)...)
 }
 
-func (l *zanzanaLogger) DebugWithContext(ctx context.Context, msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) DebugWithContext(ctx context.Context, msg string, fields ...zap.Field) {
 	l.logger.Debug(msg, zapFieldsToArgs(fields)...)
 }
 
-func (l *zanzanaLogger) InfoWithContext(ctx context.Context, msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) InfoWithContext(ctx context.Context, msg string, fields ...zap.Field) {
 	l.logger.Info(msg, zapFieldsToArgs(fields)...)
 }
 
-func (l *zanzanaLogger) WarnWithContext(ctx context.Context, msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) WarnWithContext(ctx context.Context, msg string, fields ...zap.Field) {
 	l.logger.Warn(msg, zapFieldsToArgs(fields)...)
 }
 
-func (l *zanzanaLogger) ErrorWithContext(ctx context.Context, msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) ErrorWithContext(ctx context.Context, msg string, fields ...zap.Field) {
 	l.logger.Error(msg, zapFieldsToArgs(fields)...)
 }
 
-func (l *zanzanaLogger) PanicWithContext(ctx context.Context, msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) PanicWithContext(ctx context.Context, msg string, fields ...zap.Field) {
 	l.logger.Error(msg, zapFieldsToArgs(fields)...)
 }
 
-func (l *zanzanaLogger) FatalWithContext(ctx context.Context, msg string, fields ...zap.Field) {
+func (l *ZanzanaLogger) FatalWithContext(ctx context.Context, msg string, fields ...zap.Field) {
 	l.logger.Error(msg, zapFieldsToArgs(fields)...)
 }
