+++ pkg/services/authz/zanzana/store/assets/assets.go
@@ -1,4 +1,4 @@
-package store
+package assets
 
 import "embed"
